requirements.txt
nvidia-physicsnemo
hydra-core
torch
